Added win animation and fixed camera
Camera variables are now doubles

+++ src/TileMap.java
@@ -119,7 +119,7 @@ public class TileMap
 				
 				if(c != '.')
 				{	
-					rects.add(new Rect (scale*col - Camera.x, scale*row - Camera.y,scale,scale));
+					rects.add(new Rect (scale*col - (int)Camera.x, scale*row - (int)Camera.y,scale,scale));
 				}
 			}
 		}	
@@ -250,10 +250,10 @@ public class TileMap
 	public void draw(Graphics g)
 	{
 	
-		g.drawImage(background, - Camera.x, - Camera.y, background.getWidth(null)*(scale/tileSize),background.getHeight(null)*(scale/tileSize), null);
+		g.drawImage(background, - (int)Camera.x, - (int)Camera.y, background.getWidth(null)*(scale/tileSize),background.getHeight(null)*(scale/tileSize), null);
 	
-		int c_row = Math.max(Camera.y / scale, 0);
-		int c_col = Math.max(Camera.x / scale, 0);
+		int c_row = Math.max((int)Camera.y / scale, 0);
+		int c_col = Math.max((int)Camera.x / scale, 0);
 		for(int row = c_row; row < Math.min(c_row+950/scale, map.length); row++)
 		{	
 			for(int col = c_col; col < Math.min(c_col+1500/scale, map[0].length()); col++)
@@ -262,7 +262,7 @@ public class TileMap
 				
 				if((c != '.') && ((c - 'A') < tile.length))
 				{
-				   g.drawImage(tile[c - 'A'], scale*col - Camera.x, scale*row - Camera.y, scale, scale, null); 
+				   g.drawImage(tile[c - 'A'], scale*col - (int)Camera.x, scale*row - (int)Camera.y, scale, scale, null); 
 				}
 			}
 		}
@@ -270,7 +270,7 @@ public class TileMap
 	
 	public void drawNoClipping(Graphics g)
 	{
-		g.drawImage(background, - Camera.x, - Camera.y, background.getWidth(null)*(scale/tileSize),background.getHeight(null)*(scale/tileSize), null);
+		g.drawImage(background, - (int)Camera.x, - (int)Camera.y, background.getWidth(null)*(scale/tileSize),background.getHeight(null)*(scale/tileSize), null);
 		
 		for(int row = 0; row < map.length; row++)
 		{	
@@ -280,7 +280,7 @@ public class TileMap
 				
 				if(c != '.')
 				{	
-			      g.drawImage(tile[c - 'A'], scale*col - Camera.x, scale*row - Camera.y, scale, scale, null);
+			      g.drawImage(tile[c - 'A'], scale*col - (int)Camera.x, scale*row - (int)Camera.y, scale, scale, null);
 			      rects.get(col+row).draw(g);
 				}
 			}
